don't wrap cxr, cyr with int

+++ scene/dataset_readers.py
@@ -616,8 +616,8 @@ def readPanopticmeta(datadir: str, json_path: str):
             fx, fy, cx, cy = K[0,0], K[1,1], K[0,2], K[1,2]
             FovX = focal2fov(fx, meta['w'])
             FovY = focal2fov(fy, meta['h'])
-            cxr = int((cx / meta['w']) - 0.5)
-            cyr = int((cy / meta['h']) - 0.5)
+            cxr = (cx / meta['w']) - 0.5
+            cyr = (cy / meta['h']) - 0.5
             # get R and T from w2c_list
             w2c = np.array(w2c_list, dtype=np.float32).reshape(4, 4)
             R = np.transpose(w2c[:3, :3])
